backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.9.0
pandas==2.2.2
pyarrow==16.1.0
scikit-learn==1.5.0
lightgbm==4.3.0
joblib==1.4.2
gridstatus==0.35.0
